requirements.txt
anyio==4.12.0
APScheduler==3.10.4
attrs==25.4.0
Automat==25.4.16
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.3
cssselect==1.3.0
defusedxml==0.7.1
filelock==3.20.2
h11==0.16.0
hijri-converter==2.3.2.post1
hijridate==2.3.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itemadapter==0.13.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==6.0.2
packaging==25.0
parsel==1.10.0
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyDispatcher==2.0.7
pyOpenSSL==25.3.0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2025.2
queuelib==1.8.0
requests==2.32.5
requests-file==3.0.1
Scrapy==2.14.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
tldextract==5.3.1
Twisted==25.5.0
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.2
w3lib==2.3.1
zope.interface==8.1.1
